+++ astro.config.mjs
@@ -32,7 +32,7 @@ export default defineConfig({
   integrations: [
     vue(),
     starlight({
-      title: "Mizu Design System v0.18.1",
+      title: "Mizu Design System v0.18.11",
       customCss: ["./src/styles/global.css", "@mizu/vue/style.css"],
       head: [
         {